Edit field worksheet hrd new entity dtl

+++ workspace/ims-webapp/src/main/java/th/go/excise/ims/ta/persistence/repository/TaWorksheetCondHdrRepository.java
@@ -1,10 +1,9 @@
 
 package th.go.excise.ims.ta.persistence.repository;
 
-import java.math.BigDecimal;
 import th.co.baiwa.buckwaframework.common.persistence.repository.CommonJpaCrudRepository;
 import th.go.excise.ims.ta.persistence.entity.TaWorksheetCondHdr;
 
-public interface TaWorksheetCondHdrRepository extends CommonJpaCrudRepository<TaWorksheetCondHdr, BigDecimal> {
+public interface TaWorksheetCondHdrRepository extends CommonJpaCrudRepository<TaWorksheetCondHdr, Long> {
 	public TaWorksheetCondHdr findByAnalysisNumber(String analysisNumber);
 }
